feat: Update dedicated repositories per SQL table
Product, Promotion, PromotionProduct repositories to abstract database calls to repository layer

+++ src/main/java/antigravity/repository/ProductRepository.java
@@ -31,7 +31,7 @@ public class ProductRepository {
         return Product.builder()
                 .id(rs.getInt("id"))
                 .name(rs.getString("name"))
-                .price(rs.getInt("price"))
+                .price(rs.getBigDecimal("price"))
                 .build();
     }
 }
